requirements.txt
cachetools==5.5.0
colorama==0.4.6
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyfiglet==1.0.2
requests-oauthlib==2.0.0
rsa==4.9
tabulate==0.9.0
